third_party/blink/web_tests/fast/css/class-name-no-matching-selector-expected.txt

Test that adding and removing class names with no matching selector does not cause a style recalc

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS successfullyParsed is true

TEST COMPLETE
